Project template updates

+++ prototype/manager/project_permission.php
@@ -13,13 +13,13 @@ class ManagerProjectPermission extends Manager{
 	}
     
     public final function insert($project_id, $user_email, array $permissions){
-        self::update($project_id, $user_email, $permissions, true);
+        return self::update($project_id, $user_email, $permissions, true);
     }
     
     public final function update($project_id, $user_email, array $permissions, $insert = false){
-    var_dump($user_email);
         $user = run()->manager->user->findBy(array("email"=>$user_email));
-        if($insert && $this->findBy(array("user_id"=>$user->id))){
+        $project = run()->manager->project->findBy(array("id"=>$project_id));
+        if($insert && $this->findBy(array("user_id"=>$user->id,"project_id"=>$project_id))){
 			$GLOBALS["errors"][] = (object)array(
 				"code"=>dechex(0),
 				"message"=>"User ($user_email) is already a member of this project."
@@ -44,11 +44,27 @@ class ManagerProjectPermission extends Manager{
                     parent::insert($data);
                 }
                 catch(SQLException $exc){
-                    
+                    if(strstr($exc->message(), 'Duplicate entry')){
+                    }
+                    else{
+                        $GLOBALS["errors"][] = (object)array(
+                            "code"=>dechex(0),
+                            "message"=>$exc->message()
+                        );
+                        return false;
+                    }
                 }
             }
-            else
+            else{
+                if($project->creator_id == $user->id){
+                    $GLOBALS["errors"][] = (object)array(
+                        "code"=>dechex(0),
+                        "message"=>"DELETE canceled, project owner permissions cannot be revoked."
+                    );
+                    return false;
+                }
                 parent::deleteBy($data);
+            }
         }
         return true;
     }
